tests/requirements.txt
seleniumbase
flake8
pytest-soft-assertions

